EOptInterface

Documentation for EOptInterface.

EOptInterface.decision_vars

EOptInterface.full_solution

EOptInterface.register_nlsystem

EOptInterface.register_odesystem

EOptInterface.decision_vars — Method

decision_vars(sys)

Returns the decision variables for an optimization problem formulated from a ModelingToolkit system.

EOptInterface.full_solution — Method

full_solution(model, sys)

Returns a dictionary of optimal solution values for the observed variables of an algebraic ModelingToolkit system if the JuMP model is solved.

EOptInterface.register_nlsystem — Method

register_nlsystem(model, sys, obj, ineqs)

Automatically formulates algebraic JuMP constraints and objective function from an algebraic ModelingToolkit system and user-provided constraints and objective symbolic expressions.

Arguments

model::JuMP.Model: the JuMP model

sys::ModelingToolkit.System: the ModelingToolkit system

obj::Symbolics.Num: a symbolic expression of the objective function using the ModelingToolkit system variables

ineqs::Vector{Symbolics.Num}: a vector of symbolic expressions of inequality constraints using the ModelingToolkit system variables

EOptInterface.register_odesystem — Method

register_odesystem(model, sys, tspan, tstep, integrator)

Automatically applies forward transcription and registers the discretized ODE ModelingToolkit system as algebraic JuMP constraints.

Arguments

model::JuMP.Model: the JuMP model

sys::ModelingToolkit.System: the ModelingToolkit system

tspan::Tuple{Real,Real}: the time span over which the dynamic model is simulated

tstep::Real: the time step used in the integration scheme

integrator::String: integration scheme used in discretization, "EE" for explicit Euler or "IE" for implicit Euler